Refactor out statusbar logic.
